PHPdoc annotations added, phpstan works without errors now (#465)

+++ src/Statistics/Descriptive.php
@@ -385,7 +385,14 @@ class Descriptive
      *
      * @param float[] $numbers
      *
-     * @return array (0%, Q1, Q2, Q3, 100%, IQR)
+     * @return array{
+     *     "0%":    float,
+     *     "Q1":    float,
+     *     "Q2":    float,
+     *     "Q3":    float,
+     *     "100%":  float,
+     *     "IQR":   float,
+     * }
      *
      * @throws Exception\BadDataException if the input array of numbers is empty
      */
@@ -410,8 +417,8 @@ class Descriptive
         $length = \count($numbers);
 
         if ($length % 2 == 0) {
-            $lower_half = \array_slice($numbers, 0, $length / 2);
-            $upper_half = \array_slice($numbers, $length / 2);
+            $lower_half = \array_slice($numbers, 0, (int)($length / 2));
+            $upper_half = \array_slice($numbers, (int)($length / 2));
         } else {
             $lower_half = \array_slice($numbers, 0, \intdiv($length, 2));
             $upper_half = \array_slice($numbers, \intdiv($length, 2) + 1);
@@ -456,7 +463,14 @@ class Descriptive
      *
      * @param float[] $numbers
      *
-     * @return array (0%, Q1, Q2, Q3, 100%, IQR)
+     * @return array{
+     *     "0%":    float,
+     *     "Q1":    float,
+     *     "Q2":    float,
+     *     "Q3":    float,
+     *     "100%":  float,
+     *     "IQR":   float,
+     * }
      *
      * @throws Exception\BadDataException if the input array of numbers is empty
      */
@@ -470,8 +484,8 @@ class Descriptive
         $length = \count($numbers);
 
         if ($length % 2 == 0) {
-            $lower_half = \array_slice($numbers, 0, $length / 2);
-            $upper_half = \array_slice($numbers, $length / 2);
+            $lower_half = \array_slice($numbers, 0, (int)($length / 2));
+            $upper_half = \array_slice($numbers, (int)($length / 2));
         } else {
             $lower_half = \array_slice($numbers, 0, \intdiv($length, 2));
             $upper_half = \array_slice($numbers, \intdiv($length, 2) + 1);
@@ -649,8 +663,30 @@ class Descriptive
      * @param bool    $population : true means all possible observations of the system are present;
      *                           false means a sample is used.
      *
-     * @return array [n, mean, median, mode, range, midrange, variance, sd, CV, mean_mad,
-     *                median_mad, quartiles, skewness, kurtosis, sem, ci_95, ci_99]
+     * @return array{
+     *     n:           int<0, max>,
+     *     min:         float|false,
+     *     max:         float|false,
+     *     mean:        float,
+     *     median:      float,
+     *     mode:        float[],
+     *     range:       float,
+     *     midrange:    float,
+     *     variance:    float,
+     *     sd:          float,
+     *     cv:          float,
+     *     mean_mad:    float,
+     *     median_mad:  float,
+     *     quartiles:   float[],
+     *     midhinge:    float,
+     *     skewness:    float|null,
+     *     ses:         float|null,
+     *     kurtosis:    float|null,
+     *     sek:         float|null,
+     *     sem:         float,
+     *     ci_95:       array{ci: float|null, lower_bound: float|null, upper_bound: float|null},
+     *     ci_99:       array{ci: float|null, lower_bound: float|null, upper_bound: float|null},
+     * }
      *
      * @throws Exception\OutOfBoundsException
      * @throws Exception\BadDataException
@@ -703,9 +739,15 @@ class Descriptive
      *
      * https://en.wikipedia.org/wiki/Five-number_summary
      *
-     * @param  array  $numbers
+     * @param  array<number>  $numbers
      *
-     * @return array [min, Q1, median, Q3, max]
+     * @return array{
+     *     min:     float|int|false,
+     *     Q1:      float,
+     *     median:  float,
+     *     Q3:      float,
+     *     max:     float|int|false,
+     * }
      *
      * @throws Exception\BadDataException
      */
